fix bugs

+++ src/main/java/com/dadagum/controller/CommentController.java
@@ -31,15 +31,15 @@ public class CommentController {
     @RequestMapping(value = "/production", method = RequestMethod.POST, produces = "application/json")
     @ResponseBody
     public ReturnJson<?> makeComment(@Valid Comment comment, BindingResult bindingResult, HttpSession session){
+        User user = (User) session.getAttribute("user");
+        if (user == null) return new ReturnJson<>(null, "请先登陆", false);
+        comment.setUser_id(user.getUser_id());
         System.out.println(comment);
         if (bindingResult.hasErrors()){
             List<FieldError> list = bindingResult.getFieldErrors();
             for (FieldError fieldError : list) System.out.println(fieldError.getField());
             return new ReturnJson<>(null, "传来的参数格式有错误", false);
         }
-        User user = (User) session.getAttribute("user");
-        if (user == null) return new ReturnJson<>(null, "请先登陆", false);
-        comment.setUser_id(user.getUser_id());
         List<String> result = commentService.makeComment(comment);
         return result == null ? new ReturnJson<>(new CommentDto(comment), "评论成功", true) : new ReturnJson<>(null, "评论失败", false);
     }
@@ -49,11 +49,12 @@ public class CommentController {
      * @param comment_id
      * @return
      */
-    @RequestMapping(value = "/deletion/{comment_id}", method = RequestMethod.DELETE, produces = "application/json")
+    @RequestMapping(value = "/{comment_id}", method = RequestMethod.DELETE, produces = "application/json")
     @ResponseBody
     public ReturnJson<?> deleteComment(@PathVariable int comment_id, HttpSession session){
         System.out.println(comment_id);
         User user = (User) session.getAttribute("user");
+        if (user == null) return new ReturnJson<>(null, "请先登陆", false);
         int session_id = user.getUser_id();
         return commentService.deleteComment(comment_id, session_id) ? new ReturnJson<>(null, "成功删除评论", true) : new ReturnJson<>(null, "删除评论失败", false);
     }
